# Backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Altitude Merchant API",
    "version": "1.0.0",
    "description": "Trek marketplace API, transactable by both human frontends and AI buyer agents. Every booking-affecting call is guardrail-gated (fitness, add-on cap, budget, slot availability) and written to an append-only audit trail (GET /api/v1/audit). Start at GET /api/v1/agent/catalog for a machine-readable product + capability manifest."
  },
  "servers": [{ "url": "http://localhost:5000" }],
  "paths": {
...
    "/api/v1/chat/message": {
      "post": {
        "summary": "Conversational discovery (extraction -> retrieval -> recommendation)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["message"],
            "properties": {
              "message": { "type": "string" },
              "context": { "type": "object", "properties": {
                "lastTrekId": { "type": "string" },
                "correlationId": { "type": "string" }
              } }
            }
          } } }
        },
        "responses": { "200": { "description": "One of: recommendation | clarification | booking_redirect | no_match | fallback" } }
      }
    },
...
    "/api/v1/chat/book": {
      "post": {
        "summary": "Agent-attributed reservation (source=agent). Runs the same guardrails as human checkout.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["batchId", "customerId", "customerFitnessLevel"],
            "properties": {
              "batchId": { "type": "string" },
              "customerId": { "type": "string" },
              "customerFitnessLevel": { "type": "number" },
              "addOnIds": { "type": "array", "items": { "type": "string" } },
              "maxBudget": { "type": "number" },
              "idempotencyKey": { "type": "string" },
              "correlationId": { "type": "string" }
            }
          } } }
        },
        "responses": {
          "200": { "description": "booking_success { booking, order, totalAmount } or booking_failure { text: <guardrail reason> }" }
        }
      }
    },
...
  }
}